Add revision of Code of Conduct based on feedback
To revise the Code of Conduct, we took inspiration from Ollama mistral7b (disclosed in our llm log folder). The suggestion we decided to keep was to help prevent conflicts in the first place. To do this, it suggested to foster an environment where constructive feedback is voiced early. We spun off of that inspriation and subsequently revised the 'How will we solve conflicts' section